pyproject.toml
[tool.poetry]
name = "ai_agent_agno"
version = "0.0.1"
description = "An AI agent project that uses Agno framework with data analysis tools."
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4"
pandas = "^2.0.0"
streamlit = "^1.22.0"
agno = ">=1.8.1"
duckdb = "*"
pyyaml = "^6.0.0"
openai = "^1.0.0"
python-dotenv = "^1.0.0"
numpy = "^1.24.0"
pyarrow = "^10.0.0"
fastparquet = "^0.8.0"
chardet = "^5.0.0"
sqlalchemy = "*"
plotly = "5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
